moved legendend core logic into its own component
