+++ packages/chart/src/types/Label.ts
@@ -1,0 +1,7 @@
+export type Label = {
+  datum: string
+  index: number
+  text: string
+  value: string
+  icon?: any
+}
